% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifast.R
\name{classifast_info}
\alias{classifast_info}
\title{Information about "classifast" implementation}
\usage{
classifast_info()
}
\value{
Bunch of information
}
\description{
Information about "classifast" implementation
}
