% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN.R
\name{kNN}
\alias{kNN}
\title{Apply knn algorithm}
\usage{
kNN(train, test, kfold, split, cv.iter)
}
\arguments{
\item{train}{Data frame with train data}

\item{test}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Apply knn algorithm
}
\examples{
none
}
