% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Apply logistic function to the data}
\usage{
logistic(train, test, formula)
}
\arguments{
\item{train}{Data frame with train data}

\item{test}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Apply logistic function to the data
}
\examples{
none
}
