% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{m.logistic}
\alias{m.logistic}
\title{Apply logistic function to the data}
\usage{
m.logistic(train, test, formula)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Apply logistic function to the data
}
\examples{
add(1, 1)
add(10, 1)
}
