% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifast.R
\name{predict.classifast}
\alias{predict.classifast}
\title{Predict method for the object of class "classifast"}
\usage{
\method{predict}{classifast}(x, newdata, type = NULL, cor = TRUE)
}
\arguments{
\item{x}{Object of class "classifast"}

\item{newdata}{Matrix-like objetc with new data to predict on}

\item{type}{Method for prediction. "vote" is basic stacking. "stack" constructs a meta-classifier.}

\item{cor}{TRUE by default. Prints info about the correlation of predictions of the diferent methods.}
}
\value{
Table with accuracies of different classifiers
}
\description{
Predict method for the object of class "classifast"
}
