% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForest.R
\name{RForest}
\alias{RForest}
\title{Apply Random Forest algorithm}
\usage{
RForest(train, test, kfold, split)
}
\arguments{
\item{train}{Data frame with train data}

\item{test}{A number.}
}
\value{
Random Forest model for \code{x} and \code{y}.
}
\description{
Apply Random Forest algorithm
}
\examples{
none
}
