% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifast.R
\name{summary.classifast}
\alias{summary.classifast}
\title{Summary method for the object with class "classifast"}
\usage{
\method{summary}{classifast}(x)
}
\arguments{
\item{x}{Object of class "classifast"}
}
\value{
Table with accuracies of different classifiers
}
\description{
Summary method for the object with class "classifast"
}
