% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{SVM}
\alias{SVM}
\title{Apply svm algorithm}
\usage{
SVM(train, test, kfold)
}
\arguments{
\item{train}{Data frame with train data}

\item{test}{A number.}
}
\value{
Support Vector Machine model for \code{x} and \code{y}.
}
\description{
Apply svm algorithm
}
\examples{
none
}
