% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifast.R
\name{classifast}
\alias{classifast}
\title{Add together two numbers.}
\usage{
classifast(x, y, prob, type = "simple")
}
\arguments{
\item{x}{Data frame or matrix with multivariate data with n observations (rows) and p variables (cols)}

\item{y}{A factor with the labels of the rows of x}
}
\value{
hola
}
\description{
Add together two numbers.
}
\examples{
hola
}
