% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifast.R
\name{classifast}
\alias{classifast}
\title{Compare different clasification methods on multivariante data}
\usage{
classifast(x, y, prob = 0.65, type = "simple", timing = TRUE)
}
\arguments{
\item{x}{Data frame or matrix with multivariate data with n observations (rows) and p variables (cols)}

\item{y}{A factor with the labels of the rows of x}
}
\value{
Not yet
}
\description{
Compare different clasification methods on multivariante data
}
\examples{
Not yet
}
