% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN.R
\name{kNN}
\alias{kNN}
\title{Apply kNN algorithm}
\usage{
kNN(train, test, kfold, split)
}
\arguments{
\item{train}{Data frame with train data}

\item{test}{A number.}
}
\value{
K-Nearest Neighbours model for \code{x} and \code{y}.
}
\description{
Apply kNN algorithm
}
\examples{
none
}
